requirements.txt
django == 4.0.3
djangorestframework == 3.13.1
stripe == 5.1.1
drf-yasg==1.20.0
gunicorn==20.1.0
Pillow==9.1.0
djangorestframework-simplejwt==5.1.0
django-filter==21.1
django-cors-headers==3.11.0
psycopg2-binary == 2.9.5
python-dotenv==0.20.0
